Split QSEoW thumbnail code into smaller files
